Fix Transterm typing, actually working now

+++ cryptkeeper/dependency_wrappers.py
@@ -156,8 +156,8 @@ def _read_transterm_output(input_file_name):
             new_entry["end"] = int(split_line[2])
 
         new_entry["conf"] = int(split_line[7])
-        new_entry["hairpin_score"] = int(split_line[8])
-        new_entry["tail_score"] = int(split_line[9])
+        new_entry["hairpin_score"] = float(split_line[8])
+        new_entry["tail_score"] = float(split_line[9])
 
     return entry_list
